requirements.txt
aiohttp
astroid
async-timeout
attrs
certifi
chardet
colorama
discord.py
idna
isort
lazy-object-proxy
mccabe
multidict
requests
six
snowballstemmer
tabulate
toml
urllib3
websockets
wrapt
yarl
arrow